Implemented searching as a specialized handle
git-svn-id: svn+ssh://svn.marklogic.com/project/engsvn/client-api/java/trunk@99526 62cac252-8da6-4816-9e9d-6dc37b19578c

+++ src/main/java/com/marklogic/client/config/search/SearchResults.java
@@ -9,7 +9,7 @@ package com.marklogic.client.config.search;
  */
 public interface SearchResults {
     public QueryDefinition getQueryCriteria();
-    public int getTotalResults();
+    public long getTotalResults();
 
     public SearchMetrics          getMetrics();
     public MatchDocumentSummary[] getMatchResults();
